Uses grunt-browserify to transpile and concatenate components into deployable bundle.js
