+++ client/package.json
@@ -14,10 +14,14 @@
   },
   "devDependencies": {
     "babel": "^5.8.23",
+    "babelify": "^6.3.0",
+    "browserify": "^11.2.0",
     "grunt": "^0.4.5",
     "grunt-babel": "^5.0.3",
+    "grunt-browserify": "^4.0.1",
     "grunt-contrib-concat": "^0.5.1",
     "grunt-contrib-watch": "^0.6.1",
-    "grunt-nodemon": "^0.4.0"
+    "grunt-nodemon": "^0.4.0",
+    "load-grunt-tasks": "^3.3.0"
   }
 }
